reorg vizpanel
